chore: onboarding to dynamic
