+++ apps/journeys-admin/src/components/OnboardingPanel/CreateJourneyButton/index.ts
@@ -1,0 +1,1 @@
+export { CreateJourneyButton } from './CreateJourneyButton'
