+++ apps/journeys-admin/src/components/OnboardingPanel/OnboardingList/index.ts
@@ -1,0 +1,1 @@
+export { OnboardingList } from './OnboardingList'
