+++ apps/journeys-admin/src/components/OnboardingPanel/OnboardingListLoading/OnboardingListLoading.tsx
@@ -1,0 +1,13 @@
+import { ReactElement } from 'react'
+
+import { MediaListItem } from '../../MediaListItem'
+
+export function OnboardingListLoading(): ReactElement {
+  return (
+    <>
+      <MediaListItem loading />
+      <MediaListItem loading />
+      <MediaListItem loading />
+    </>
+  )
+}
