+++ apps/journeys-admin/src/components/OnboardingPanel/OnboardingListLoading/index.ts
@@ -1,0 +1,1 @@
+export { OnboardingListLoading } from './OnboardingListLoading'
