+++ apps/journeys-admin/src/components/OnboardingPanel/index.ts
@@ -1,0 +1,1 @@
+export { OnboardingPanel } from './OnboardingPanel'
