--- apps/journeys-admin/src/components/OnboardingPanelContent/index.ts
@@ -1,1 +1,0 @@
-export { OnboardingPanelContent } from './OnboardingPanelContent'
